feat(security): redact secrets from tool results before LLM context
Extract secret registry into src/redact.ts so it can be shared between
the log layer and the tool layer without circular dependencies.

Wrap every tool's execute() in withRedaction() (tools/index.ts) so that:
- Text content blocks returned to the LLM are redacted
- Error messages thrown from tools (e.g. bash non-zero exit with output)
  are also redacted before reaching the LLM context window

This closes the remaining leak path where a secret could silently survive
in context.jsonl even though it was already scrubbed from logs.
